+++ frontend/src/components/image/ImageCard.tsx
