+++ frontend/src/components/image/ImageViewer.tsx
