+++ frontend/src/components/login/NoAuth.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import {Navigate} from "react-router-dom";
-import useAuth from "../hooks/useAuth";
+import useAuth from "../../hooks/useAuth";
 
 export default function NoAuth (
     {
