+++ frontend/src/components/patient/GenderRadioGroup.tsx
@@ -5,7 +5,7 @@ import RadioGroup from '@mui/material/RadioGroup';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import FormControl from '@mui/material/FormControl';
 import FormLabel from '@mui/material/FormLabel';
-import Patient from "../types/Patient";
+import Patient from "../../types/Patient";
 
 export default function GenderRadioButtonsGroup({
     patient,
